% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{tl_metadata}
\alias{tl_metadata}
\title{Dataset metadata}
\usage{
tl_metadata(dataset, locale = NULL)
}
\arguments{
\item{dataset}{A `topolab_dataset`.}

\item{locale}{Optional locale ("en"/"nl").}
}
\value{
Parsed dataset metadata (list).
}
\description{
Dataset metadata
}
